1.python/django_vue/api/requirements.txt
Django==2.2.5
psycopg2
djangorestframework
django-cors-headers